backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
slowapi==0.1.5
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
